+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Handle user interactions for overdue search results
 - Implement overdue search results adapter
 - Handle overdue search load states
+- Search overdue patients, when the search history item is clicked
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ## 2022-06-20-8300